+++ renderer/Sound.h
@@ -13,8 +13,8 @@ namespace Transmission {
 		static Sound* Sound::createSound( Window* window );
 
 		int LoadSoundFile( char* file );
-		virtual bool PlaySoundAt( int index, bool loop ) = 0;
-		virtual bool StopSoundAt( int index ) = 0;
+		virtual bool PlaySoundAt( size_t index, bool loop ) = 0;
+		virtual bool StopSoundAt( size_t index ) = 0;
 
 	private:
 		virtual int LoadWaveFile( char* file ) = 0;
